requirements.txt
streamlit
faster-whisper
requests
google-generativeai
python-dotenv
python-dotenv
